refactor: move command bar tests
